Centralize detection of ffmpeg executables
In some distributions (e.g. NixOS) ffmpeg is not necessarily in PATH. Unifying
detection of the ffmpeg and ffprobe executables makes it easier for packagers
to patch the package to accomodate such situations.
