+++ stempeg/ffmpeg.py
@@ -1,0 +1,30 @@
+FFMPEG_PATH = None
+FFPROBE_PATH = None
+
+def find_cmd(cmd):
+    try:
+        from shutil import which
+        return which(cmd)
+    except ImportError:
+        import os
+        for path in os.environ["PATH"].split(os.pathsep):
+            if os.access(os.path.join(path, cmd), os.X_OK):
+                return path
+
+    return None
+
+def ffmpeg_and_ffprobe_exists():
+    global FFMPEG_PATH, FFPROBE_PATH
+    if FFMPEG_PATH is None:
+        FFMPEG_PATH = find_cmd("ffmpeg")
+
+    if FFPROBE_PATH is None:
+        FFPROBE_PATH = find_cmd("ffprobe")
+
+    return FFMPEG_PATH is not None and FFPROBE_PATH is not None
+
+if not ffmpeg_and_ffprobe_exists():
+    raise RuntimeError('ffmpeg or ffprobe could not be found! '
+                       'Please install them before using stempeg. '
+                       'See: https://github.com/faroit/stempeg')
+
